Refactor: Consolidate ApplicationsFetcher functionality into SBMLPage
